% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{community}
\alias{community}
\title{Determine community spectrum}
\usage{
community(p, sim)
}
\arguments{
\item{p}{matrix of population densities (N x Ns)}

\item{sim}{Sim object}
}
\value{
A vector of community population densities at points \code{xa}
}
\description{
Adds together the population density of all species
wrapped around assuming periodicity in species size.
Used in \code{\link{evolve_cel_pop}}.
This returns the $\tilde{p}_c$ from the vignette that
is constant in the steady-state.
}

