% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{community_spectrum}
\alias{community_spectrum}
\title{Determine community spectrum}
\usage{
community_spectrum(p, r)
}
\arguments{
\item{p}{matrix of population densities (N x Ns)}

\item{r}{PlanktonParams object}
}
\value{
A vector of community population densities at points \code{xa}
}
\description{
Adds together the population density of all species
wrapped around assuming periodicity in species size
Used in \code{\link{evolve_cel_pop}}
}

