% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{evolve_cell_pop}
\alias{evolve_cell_pop}
\title{Evolve cell population density using the population balance equation}
\usage{
evolve_cell_pop(r)
}
\arguments{
\item{r}{PlanktonSim object}
}
\value{
list of two elements:
    Nt x N x Ns  array of population densities
    vector of length Nt containing the nutrient densities
}
\description{
see eq.(2.10)
}

