% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{fourier_interpolate}
\alias{fourier_interpolate}
\title{Perform a Fourier interpolation}
\usage{
fourier_interpolate(p, n)
}
\arguments{
\item{p}{vector of values of function at equally spaced steps, excluding
the right endpoint}

\item{n}{desired length of output vector}
}
\value{
Vector of n values of Fourier interpolation at n equally
    spaced points, excluding the right endpoints.
}
\description{
Takes a discretisation at any number of equally-spaced points, performs
a Fast Fourier Transform on it and then does a slow inverse Fourier
transform evaluated at the desired number of equally-spaced points.
}

