% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{get_community}
\alias{get_community}
\title{Get community spectrum}
\usage{
get_community(sim)
}
\arguments{
\item{sim}{Sim object}
}
\value{
matrix Nt x Na
}
\description{
This produces $\tilde{p}_c(t, w)$ as defined in the vignette.
In the steady state this should be constant in w.
}

