% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{getGrid}
\alias{getGrid}
\title{Extract grid from simulation}
\usage{
getGrid(sim)
}
\arguments{
\item{sim}{Object of class Sim}
}
\value{
Object of class Grid
}
\description{
This can be achieved just as easy with as(). I am providing this function
only lest I might forget how to use as()
}

