% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{getParams}
\alias{getParams}
\title{Extract parameters from simulation}
\usage{
getParams(sim)
}
\arguments{
\item{sim}{Object of class Sim}
}
\value{
Object of class Grid
}
\description{
Extract parameters from simulation
}

