% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\docType{class}
\name{Grid-class}
\alias{Grid-class}
\alias{plot,Grid-method}
\alias{show,Grid-method}
\title{An S4 class to represent a 3d grid on which to do a simulation}
\usage{
\S4method{plot}{Grid}(x, y, ...)

\S4method{show}{Grid}(object)
}
\description{
An S4 class to represent a 3d grid on which to do a simulation
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Draws a picture of the grid

\item \code{show}: Show very short description of object
}}
\section{Slots}{

\describe{
\item{\code{N}}{Number of steps within-species}

\item{\code{Ns}}{Number of species}

\item{\code{SpeciesSpacing}}{Number of species per length of species}

\item{\code{x}}{Log sizes within a species.}

\item{\code{dx}}{log step size within a species.}

\item{\code{w}}{Sizes within a species.}

\item{\code{ds}}{Number of steps between species.}

\item{\code{dxs}}{Log distance between species.}

\item{\code{xs}}{Log maximum sizes of species.}

\item{\code{ws}}{Maximum sizes of species.}

\item{\code{xa}}{Log sizes within community.}

\item{\code{Na}}{Number of steps in community spectrum.}

\item{\code{xal}}{Log sizes in community spectrum before periodic wrapping.}

\item{\code{Nal}}{Number of steps in community spectrum before periodic wrapping.}

\item{\code{wsmgamma}}{\code{ws^(-gamma)}}

\item{\code{walgamma}}{\code{wal^(gamma)}}

\item{\code{tmax}}{Time until which to run simulation. If the vector \code{t} of
times is supplied than \code{tmax} is set from that.}

\item{\code{Nt}}{Number of time steps at which to return population density.
If the vector \code{t} is supplied than \code{Nt} is set from that.}

\item{\code{t}}{vector of \code{Nt} time steps between 0 and \code{tmax}. If this
is not supplied then the default is a vector of \code{Nt} equally-spaced
time points between 0 and \code{tmax}.}
}}

