% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{make_p0}
\alias{make_p0}
\title{Make initial population density}
\usage{
make_p0(r, psi, pp)
}
\arguments{
\item{r}{Object of type Grid. An object of type Sim will be coerced to type
Grid.}

\item{psi}{Vector giving unnormalised single-species steady
state solution. If this does not have length N it will
be subsampled with Fourier interpolation. Default \code{r@psiBar}.}

\item{pp}{Vector of length r@Ns giving the relative normalisation of the
population densities. If this does not have length N it will
be subsampled with Fourier interpolation. Default constant.}
}
\value{
Matrix of dimension N x Ns
}
\description{
Produces an N x Ns matrix of population densities that can be used as
initial condition for a simulation. It assumes that all species should
start with the same initial profile \code{psi} but with possibly different
normalisations \code{pp}.
}
\details{
The initial population density will be normalised so that the
community spectrum is 1.
}

