% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{class}
\name{Params-class}
\alias{Params-class}
\alias{plot,Params-method}
\alias{show,Params-method}
\title{An S4 class to represent the parameters of a plankton model}
\usage{
\S4method{plot}{Params}(x, y)

\S4method{show}{Params}(object)
}
\description{
An S4 class to represent the parameters of a plankton model
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot the steady-state solution

\item \code{show}: Show very short description of object
}}
\section{Slots}{

\describe{
\item{\code{xi}}{Exponent for allometric scaling of mortality. Default 0.15}

\item{\code{nu}}{Exponent in scaling of predation kernel. Default 0.85}

\item{\code{gamma}}{Exponent in across-species steady-state power-law.
Equal to 1+nu+xi.}

\item{\code{a_inf}}{Nutrient consumption rate when unlimited nutrient supply.
Default 2}

\item{\code{rr}}{Half saturation level at which nutrient consumption is half of the
maximal value \code{a_inf}.}

\item{\code{rho_0}}{Nutrient replenishment rate when nutrient is low.}

\item{\code{Nu_0}}{Nutrient carrying capacity in absence of consumption.}

\item{\code{w_th}}{Threshold size for duplication. Cells smaller than \code{w_th}
can not duplicate.}

\item{\code{delta_q}}{Width of offspring size distribution. Daughter cells can have
a weight that lies between \code{(1-delta_q)/2} and \code{(1+delta_q)/2}
of their parent's weight.}

\item{\code{k_0}}{Overall coefficient of duplication rate function.}

\item{\code{ke}}{Exponent in duplication rate function.}

\item{\code{alpha}}{Allometric exponent of nutrient intake rate.}

\item{\code{beta}}{Allometric exponent of metabolic loss rate.}

\item{\code{b}}{Coefficient of metabolic loss rate.}

\item{\code{m}}{Coefficient of death rate.}

\item{\code{epsilon}}{Conversion efficiency of prey mass into predator mass.}

\item{\code{s0}}{Coefficient of predation kernel.}

\item{\code{beta_p}}{Center of predation kernel.}

\item{\code{delta_p}}{Width of predation kernel.}

\item{\code{a}}{Function \code{a(Nu)} giving nutrient intake rate coefficient as a
function of nutrient concentration.}

\item{\code{dNu}}{Function \code{dNu(w, Nu, psi, r)} giving the nutrient growth
rate for a cell of weight \code{w} as a function of nutrient concentration
\code{Nu} and plankton population \code{psi}.}

\item{\code{k}}{Function \code{k(w)} giving duplication rate for a cell of weight
\code{w}}

\item{\code{g}}{Function \code{g(w, Nu)} giving the growth due to resource
consumption for a cell of weight \code{w} as a function of resource
concentration \code{Nu}.}

\item{\code{s}}{Function \code{s(x)} giving the predation kernel as a function of
the logarithm \code{x} of the predator/prey mass ratio. The predation kernel
needs to be multiplied by the allometric scaling factor and the predator
and prey densities to give the actual predation rate.}

\item{\code{L}}{Length of single-cell size distribution in log-size.}

\item{\code{wBar}}{Weights at which the steady-state population is given.}

\item{\code{psiBar}}{Population density at steady state.}

\item{\code{NuBar}}{Nutrient concentration at steady state.}
}}

