% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{Params}
\alias{Params}
\title{Create and initialise Params object}
\usage{
Params(...)
}
\arguments{
\item{...}{Arguments that will be passed to \code{new()}.}
}
\description{
Create and initialise Params object
}

