% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\docType{methods}
\name{plot3d}
\alias{plot3d}
\alias{plot3d,Sim,missing,missing-method}
\alias{plot3d,Sim,missing,numeric-method}
\alias{plot3d,Sim,numeric,missing-method}
\title{Produce 3 dimenstional plot of the simulation result}
\usage{
plot3d(r, t, xs)

\S4method{plot3d}{Sim,numeric,missing}(r, t)

\S4method{plot3d}{Sim,missing,numeric}(r, xs)

\S4method{plot3d}{Sim,missing,missing}(r)
}
\arguments{
\item{r}{Object of type Sim}

\item{t}{Time at which to plot. If the time is not one of the values in r@t,
choose next larger time contained in r@t or, if there is none, choose
largest time in r@t.}

\item{xs}{Log size of species to plot. If there is no species at this size,
choose the next larger species, or if there is none, the largest species.}
}
\description{
Produce 3 dimenstional plot of the simulation result
}
\section{Methods (by class)}{
\itemize{
\item \code{r = Sim,t = numeric,xs = missing}: 3d plot of population density against x and xa at given time

\item \code{r = Sim,t = missing,xs = numeric}: 3d plot of population density of given species against t and x

\item \code{r = Sim,t = missing,xs = missing}: 3d plot of community density against x and xa
}}

