% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{plot3d_community}
\alias{plot3d_community}
\title{Plot community spectrum against time and size}
\usage{
plot3d_community(sim)
}
\arguments{
\item{sim}{Sim object}
}
\description{
This plots $\tilde{p}_c(t, w)$ as defined in the vignette.
}

