% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{plot_community}
\alias{plot_community}
\title{Plot community spectrum against size at one time}
\usage{
plot_community(sim, t = NULL)
}
\arguments{
\item{sim}{Sim object}

\item{t}{Time at which to plot. If the value is not available
at that time, the last earlier time is used. Default: latest available time.}
}
\description{
This plots \eqn{\tilde{p}_c(t, w)} as defined in the vignette.
}

