% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{s_moment}
\alias{s_moment}
\title{Calculate moment of predation kernel}
\usage{
s_moment(r, lambda)
}
\arguments{
\item{r}{Object of type Params}

\item{lambda}{The exponent}
}
\value{
The value of the lambda-th moment
}
\description{
The integral over s(x) * exp(lambda x)
}

