% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{steady_state}
\alias{steady_state}
\title{Determine steady state cell size distribution}
\usage{
steady_state(r)
}
\arguments{
\item{r}{Object of class PlanktonParams}
}
\value{
List containing:
  \enumerate{
    \item vector of steady-state population densities \eqn{\psi}
    \item nutrient concentration at steady state
  }
}
\description{
Using the analytic formulae in eqs.(4.16) to (4.21) of the paper
}

