% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{evolve_cell_pop}
\alias{evolve_cell_pop}
\title{Evolve cell population density using the population balance equation}
\usage{
evolve_cell_pop(p0, Nu0, r)
}
\arguments{
\item{p0}{N x Ns matrix or vector with initial population for all Ns species
or a vector of length N with initial population for one species that is then
replicated over all species. If missing, the steady-state solution is used
and replicated over all species.}

\item{Nu0}{Initial resource concentration. If missing, the steady-state value
is used}

\item{r}{Sim object}
}
\value{
list of two elements:
    Nt x N x Ns  array of population densities
    vector of length Nt containing the nutrient densities
}
\description{
see eq.(2.10)
}

