% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\docType{class}
\name{Sim-class}
\alias{Sim-class}
\alias{plot,Sim-method}
\alias{show,Sim-method}
\alias{summary,Sim-method}
\title{An S4 class to represent the simulation of a plankton model}
\usage{
\S4method{plot}{Sim}(x, y, t = NULL, s = 1, xlog = FALSE)

\S4method{show}{Sim}(object)

\S4method{summary}{Sim}(object)
}
\arguments{
\item{t}{Time at which to plot. If the value is not available
at that time, the next larger time is used. Default: latest available time.}

\item{s}{Index of species to plot. Default 1.}

\item{sim}{Sim object}
}
\description{
An S4 class to represent the simulation of a plankton model
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot the solution for one species at one time

\item \code{show}: Show very short description of object

\item \code{summary}: List grid and model parameter values
}}
\section{Slots}{

\describe{
\item{\code{p}}{Array (Nt x N x Ns) of simulated population densities.}

\item{\code{Nu}}{Vector of simulated nutrient concentrations.}

\item{\code{rho_0}}{Nutrient replenishment rate when nutrient is low.}

\item{\code{Nu_0}}{Nutrient carrying capacity in absence of consumption.}

\item{\code{dNu}}{Function giving nutrient replenishment rate.}
}}

